Fix a bunch of MSVC compiler warnings and errors.
Most notably, we need to actually fix a bug in MSVC that doesn't know how to properly instantiate
enums in partial template specialization. There are more details outlined here:

http://stackoverflow.com/questions/15466594/why-does-msvc-fail-to-compile-this-template-function

The fix in this commit closes #10

Also in this commit is a hacky way to allow GL defines. Apparently "LoadImage" is defined as a
macro even with WIN32_LEAN_AND_MEAN. This means that we have to #undef the code that includes
it, meaning that we also need to make sure not to actually mix GLDefines.h with any file that needs
to use the functions from Windows.h

+++ Base/test/TestMatrix.cpp
@@ -53,7 +53,7 @@
 #include "gtest/gtest.h"
 #include "MatrixBase.h"
 
-static const float kEpsilon = 1e-6;
+static const float kEpsilon = 1e-6f;
 
 TEST(MatrixBase, Constructors) {
   FasTC::MatrixBase<float, 3, 4> m3f;
@@ -123,7 +123,7 @@ TEST(MatrixBase, PointerConversion) {
 }
 
 TEST(MatrixBase, CastVector) {
-  srand(time(NULL));
+  srand(static_cast<unsigned>(time(NULL)));
 
   FasTC::MatrixBase<float, 3, 2> v3f;
   for(int i = 0; i < 6; i++) {
